pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "price-tracker"
version = "0.1.0"
description = "A web scraper to track product prices"
authors = ["Adrien Lynch <adriienlynch@gmail.com>"]

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.8.1" 
python-dotenv = ">1.0.0"

requests = ">2.28.1"
beautifulsoup4 = ">4.11.1"
lxml = ">4.9.0"
selenium = ">4.15.0"
webdriver-manager = ">4.0.0"
cryptography = ">41.0.0"
pandas = ">1.5.1"
twilio = ">7.12.0"
flake8 = "^7.1.1"

[tool.poetry.dev-dependencies]
# Development dependencies here, e.g. testing frameworks
pytest = ">7.0"
